Move common class to the upper level
