+++ modules/core/src/test/groovy/config_interfaces/ConfigInterfacesTest.groovy
@@ -17,7 +17,7 @@
 package config_interfaces
 
 import test_support.addon1.TestAddon1Configuration
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import test_support.app.TestAppConfiguration
 import test_support.app.TestConfig
 import io.jmix.core.ConfigInterfaces
