+++ modules/core/src/test/groovy/datatypes/DatatypeRegistryTest.groovy
@@ -16,7 +16,7 @@
 
 package datatypes
 
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import io.jmix.core.JmixCoreConfiguration
 import io.jmix.core.metamodel.datatypes.DatatypeRegistry
 import io.jmix.core.metamodel.datatypes.impl.*
