+++ modules/core/src/test/groovy/entity_enhancing/EntityEnhancingTest.groovy
@@ -18,7 +18,7 @@ package entity_enhancing
 
 import test_support.addon1.TestAddon1Configuration
 import test_support.addon1.entity.TestAddon1Entity
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import test_support.app.TestAppConfiguration
 import test_support.app.entity.Pet
 import io.jmix.core.JmixCoreConfiguration
