+++ modules/core/src/test/groovy/entity_scanning/EntitiesScannerMultiConfigTest.groovy
@@ -17,7 +17,7 @@
 package entity_scanning
 
 import test_support.addon1.TestAddon1Configuration
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import io.jmix.core.JmixCoreConfiguration
 import io.jmix.core.impl.scanning.EntitiesScanner
 import org.springframework.beans.factory.annotation.Autowired
