+++ modules/core/src/test/groovy/jmix_modules/JmixModulesTest.groovy
@@ -17,7 +17,7 @@
 package jmix_modules
 
 import test_support.addon1.TestAddon1Configuration
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import test_support.app.TestAppConfiguration
 import test_support.app.TestBean
 import io.jmix.core.JmixCoreConfiguration
