+++ modules/core/src/test/groovy/messages/MessageToolsTest.groovy
@@ -17,7 +17,7 @@
 package messages
 
 import test_support.addon1.TestAddon1Configuration
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import test_support.app.TestAppConfiguration
 import test_support.app.entity.Pet
 import io.jmix.core.JmixCoreConfiguration
