+++ modules/core/src/test/groovy/metadata/MetadataToolsTest.groovy
@@ -17,7 +17,7 @@
 package metadata
 
 import test_support.addon1.TestAddon1Configuration
-import test_support.app.AppContextTestExecutionListener
+import test_support.AppContextTestExecutionListener
 import test_support.app.TestAppConfiguration
 import test_support.app.entity.Owner
 import test_support.app.entity.Pet
